Added lib/inflateMeta, renamed helpers to lib/inflate
