+++ lib/inflate.js
@@ -1,23 +1,8 @@
-// TODO - split this up into one-function-per-file
-const titleize = require('underscore.string/titleize')
-const underscored = require('underscore.string/underscored')
-const classify = require('underscore.string/classify')
-const pluralize = require('pluralize')
+const { inflateMeta } = require('./inflateMeta')
 const clone = require('lodash/clone')
 const cloneDeep = require('lodash/cloneDeep')
 
 // // // //
-
-function inflateMeta(label) {
-  return {
-    label: titleize(label),
-    label_plural: pluralize(titleize(label)),
-    identifier: underscored(label),
-    identifier_plural: underscored(pluralize(label)),
-    class_name: classify(titleize(label)),
-    class_name_plural: pluralize(classify(titleize(label)))
-  }
-}
 
 function inflateRelation({ schemas, relation }) {
   // Clones the base attributes from the relation
@@ -53,6 +38,7 @@ function inflateSchema({ schema, schemas }) {
 // inflate
 // Fomats the build parameters for the generator
 // Mostly adds some additional metadata to each relation to simplify template rendering
+// TODO - this may want to define the app.identifier attribute (rather than relying on the client)
 function inflate({ app }) {
 
     // Iterates over each schema
@@ -102,6 +88,5 @@ function inflate({ app }) {
 module.exports = {
   inflateSchema,
   inflateRelation,
-  inflateMeta,
   inflate
 }
